Almost finished Pipe (missing support for |&) and Fixed some bugs

+++ signals.cpp
@@ -2,6 +2,8 @@
 #include <signal.h>
 #include "signals.h"
 #include "Commands.h"
+#include <unistd.h>
+
 
 using namespace std;
 
@@ -10,27 +12,66 @@ bool sigINTOn = false;
 
 void ctrlCHandler(int sig_num) {
     cout << "smash: got ctrl-C" << endl;
-    //TODO: if there is no process in the foreground stop here!
-
-    //TODO: for pipes - don't send SIGKILL, send SIGINT and pipe
-    // handler will stop it's sons and then kill pipe, after that
-    // rememeber to set isForegroundPipe to false after handling
-    kill(foregroundPid, SIGKILL);
+    if (foregroundPid == 0){
+        return;
+    }
+    sigINTOn = true;
+    if(isForegroundPipe){
+        kill(foregroundPid, SIGINT);
+    }
+    else {
+        kill(foregroundPid, SIGKILL);
+    }
     cout << "smash: process " << foregroundPid << " was killed" << endl;
 }
 
 void ctrlZHandler(int sig_num) {
     cout << "smash: got ctrl-Z" << endl;
-    //TODO: if there is no process in the foreground stop here!
-
-    //TODO: for pipes - don't send SIGSTOP, send SIGSTP and pipe
-    // handler will stop it's sons and then STOP pipe, after that
-    // rememeber to set isForegroundPipe to false after handling
-    kill(foregroundPid, SIGSTOP);
+    if (foregroundPid == 0){
+        return;
+    }
+    sigSTPOn = true;
+    if(isForegroundPipe){
+        kill(foregroundPid, SIGTSTP);
+    }
+    else {
+        kill(foregroundPid, SIGSTOP);
+    }
     cout << "smash: process " << foregroundPid << " was stopped" << endl;
 }
 
-//TODO: do we need a handler to SIGCONT for a pipe?
+void pipeCtrlCHandler(int sig_num){
+    sigINTOn = true; //pay attention, this is copy of the global sigSTPOn of the smash process, it's the same one!!
+    if (pipeFirstCmdPid != NOT_FORKED){
+        kill(pipeFirstCmdPid, SIGKILL);
+    }
+    if(pipeSecondCmdPid != NOT_FORKED){
+        kill(pipeSecondCmdPid, SIGKILL);
+    }
+}
+
+void pipeCtrlZHandler(int sig_num){
+    if (pipeFirstCmdPid != NOT_FORKED){
+        kill(pipeFirstCmdPid, SIGSTOP);
+    }
+    if(pipeSecondCmdPid != NOT_FORKED){
+        kill(pipeSecondCmdPid, SIGSTOP);
+    }
+    kill(getpid(), SIGSTOP);
+}
+
+void pipeSigcontHandler(int sig_num){
+    if (pipeFirstCmdPid != NOT_FORKED){
+        kill(pipeFirstCmdPid, SIGCONT);
+    }
+    if(pipeSecondCmdPid != NOT_FORKED){
+        kill(pipeSecondCmdPid, SIGCONT);
+    }
+    kill(getpid(), SIGCONT);
+    /*after this, pipe should restart the waitpid in the pipe execute
+     * to wait for his sons that have been already continued until they finish
+     * or until the next signal interrupt*/
+}
 
 void alarmHandler(int sig_num) {
     // TODO: Add your implementation
